Refactors config implementation: Use java.util.Properties directly now
